+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: RLUSD on Ethereum and Ripple networks
+
 ## 4.37.3 (2025-02-28)
 
 - changed: (Abstract) Ignore internal transactions